requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
djangorestframework==3.9.4
httpie==1.0.2
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
